dnsdist: Move the rings declaration to a separate header file
