+++ pdns/dnsdist-lua-inspection.cc
@@ -22,6 +22,7 @@
 #include "dnsdist.hh"
 #include "dnsdist-lua.hh"
 #include "dnsdist-dynblocks.hh"
+#include "dnsdist-rings.hh"
 
 #include "statnode.hh"
