+++ pdns/dnsdist-lua.cc
@@ -31,6 +31,7 @@
 #include "dnsdist.hh"
 #include "dnsdist-console.hh"
 #include "dnsdist-lua.hh"
+#include "dnsdist-rings.hh"
 
 #include "base64.hh"
 #include "dnswriter.hh"
