+++ pdns/dnsdist.cc
@@ -46,6 +46,7 @@
 #include "dnsdist-console.hh"
 #include "dnsdist-ecs.hh"
 #include "dnsdist-lua.hh"
+#include "dnsdist-rings.hh"
 
 #include "base64.hh"
 #include "delaypipe.hh"
