+++ pdns/dnsdistdist/Makefile.am
@@ -98,7 +98,7 @@ dnsdist_SOURCES = \
 	dnsdist-lua-rules.cc \
 	dnsdist-lua-vars.cc \
 	dnsdist-protobuf.cc dnsdist-protobuf.hh \
-	dnsdist-rings.cc \
+	dnsdist-rings.cc dnsdist-rings.hh \
 	dnsdist-snmp.cc dnsdist-snmp.hh \
 	dnsdist-tcp.cc \
 	dnsdist-web.cc \
