+++ pdns/dnsdistdist/dnsdist-rings.hh
@@ -1,0 +1,1 @@
+../dnsdist-rings.hh
